Backend: Implemented GraphQL API
GraphQL is integrated into the application and is found in the src/API/GraphQL folder.
The folder contains a file name of route.js and two folders named Model and Query.

The file route.js defines all routes the GraphQL API offers to a client.
ROOTQUERY contains all routes using a query retrieving data. If a route would have used a GET-Request, this is where the route and the query belongs.
MUTATION contains all routes using  a query updating or deleting data.
 If a route would have used a PUT- or DELETE-Request, than this is where the route and the query belongs.

The Query folder contains code implementing the queries used to retrieve or update data with GraphQL.
The Model folder contains definitions of the models used by the application. The models are representation of different objects, like the achievement object found in the folder src/BusinessLogic/GameElements.

The file index.js in folder API only exports the GraphQL Schema which is imported by the file internal.js.

The file server.js imports GraphQL Schema from the file internal.js and graphqlHTTP method which is used by the application.
The application defines a middelware for GraphQL which allows user to access the GraphQL GUI from their browser.

The file internal.js exports all GraphQL functionality from the API folder and makes it available to the rest of the application to import.
Spaces added between each foleder name in internal.js file.

Added GraphQL and express-graphql as dependencies for the projects package.json file.

Deleted the api.js file and user_endpoints.js files from the folder Routes because the application now uses GraphQL as API.