Backend: Dummy Data, Test, Config, and Database
Created a new yaml file for development. Contains environment variables.
The file is used when running the application in development mode and NODE_ENV is set to development.

Added new scripts to the package.json file.
One script for to run application in development mode and one for production.
Installed dotenv-yaml dependency used to read yaml files by the config.js file.

Updated the config.js file to switch yaml file and environment variables  depending on if the application is launched in production or development mode.

Added new routes to GraphQL.
One route for getting users by id, one for deleting a user, and one for updating a user.

Implemented new queries for GraphQL. New queries gets user by, creates a user in the database, delete a user in the database, and update a user in the database.

The connection.js file now checks if the application is in development or production mode and connects to the database used in either of the modes.

The query.js file implements functions retreiving, deleting, updating, and filtering data from database in development mode.

Created a set with dummy data and a backup with users in the application.
The dummy data is located in src/Database folder.
The dummy data for users are created by using mockaroo.com to generate fake data.

Added a data set to use for 2 way testing of profile functionality.
The test set is found in backend/test/DataSet folder.

Created files to write unit tests for access token, profile, and sign in functionality of the application.