Backend: Hash and Compare Password
Changes made to the file hashing.js in the src/Security folder.
These changes makes the application able to secure passwords with hashing and to sign in users by comparing passwords.

Implemented a hash password function taking a plain text password as argument.
Returns the hashed result.
Function must be used when a user register, changes their password or has forgotten their password.

Implemented a compare password function taking a plain text user password and compares it with the hashed password stored in database.
Returns an object containing a bool stateing if the passwords matched or not.