Backend: Access token and Refresh token
Changes made to file access_token.js in the src/Authorization/AccessToken folder.
Changes makes the application able to create an access token and refresh_token for the user when they sign in.
The access token is used to check if the user is signed in to the application and to authorize use of API endpoints require a signed in user.
The refresh token is used to refresh an expired access token and is the only token to be stored in database.
Refresh token signs out the user if the refresh token is deleted from database.

Implements a create access token function taking a user as argument and uses the user data to sign an acces token along with a secret.
The function is used by the sign in logic and when a refresh token is authenticated.

Implements an authenticate access token function taking an express next function as an argument, which passes the control to the next function in chain.
Authentication of access token is performed by checking the authorization header of a request and decrypting the token. Decryption returns a user as long as the token is not invalid.

Implements a create refresh token function taking in a user as argument.
Uses the user data and a secret to create a refresh token which is stored in database.
The refresh token is stored in database until a user is signed out to allow the application to use the refresh token to refresh an expired access token.
The create refresh token is used by the sign in logic.

Implements an authenticate refresh token function taking a refresh token as an argument.
Fetches refresh tokens from database and verifies if it exists. If the refresh token is not found, an response of 403 is returned. If it is found the token gets verified and returns a new access token.

Implements a delete refresh token from database function taking a refresh token as an argument.
Fetches all refresh tokens from database and verifies if the user is still signed in.
If user is signed in the refresh token is deleted from database and a response of 200 is returned to the user.

Implements a store refresh token in database function taking a refresh token as an argument which should be stored in database.
Fetches all refresh tokens from database and verifies if the refresh token passed as argument already exists.
Existing tokens is not stored again and returns a response of 200.
Non-existing refresh token updates the database with a new refresh token and returns a response of 200.

Updated development.yaml with a new environment variable storing the table name in database for refresh tokens.

Updated the config.js file to provide a refresh token table key/value pair for the application to use.