Backend: Fixed Logic Flaws
Changes made to the files access_token.js and query.js in folder src/Authorization/AccessToken and src/Database/Query respectively.

Added missing try/catch clauses to functions in access_token.js.
The error caught are now re thrown to allow the caller function to handle the error and send a response.
Added missing variables in authenticate access token function.

Updated the has data entry function in query.js to check if the it is passed a list or an object as an argument.
The check is to handle the case in development mode where the refresh_token table in dummy data is a list with strings and the user table is a list of objects.

Added development.yaml and app.yaml files to gitignore.