Backend: Sign In
Changes made to file sign_in.js in src/Authentication folder.
Changes makes the application able to sign in a user and return an access token, refresh token and the data used stored about the user mapped to the username signed in with.

Implements a sign in function taking username and password as argument and returns a promise.
First it verifies that a user with the username passed in as an argument exist.
Then it compares the user's stored hashed password with the password passed in as an argument.
If the passwords match, an access token and refresh token is created.
The function returns an object containing the user data of the signed in user, a access token and a refresh token.

Added a new environment variable to development .yaml storing the name of the user table in the dummy data set.

Added a new key/value pair to the config.js file for user table name. Also added a key/value pair for response type, defining a response to either success or error.