Backend: GraphQL Sign In Route
Changes made to files query.js, model.js, and route.js in src/API/GraphQL.
The changes updates the model, query, and route of GraphQL to provide a Sign In endpoint.
The endpoint makes the application able to sign in users from the frontend.

Added a new sign in model to query.js. The model states how GrapQL should respond when a user sign in to the application.
The sign in model uses a general response model and a data field. The response model populating the response with a message, status, and type field.
The message field allows frontend to provide user feedback, regardless if it is an error message or a regular message.
The status field contains the http status code of the response.
The type field contains what type of response it is. It could be an error or a success response.
The data field is populated by a sign in data model, which have an access token, refresh token, and user field.
The access token and refresh token field is a string representing the access token and refresh token respectively.
The user field is populated by a user model, which is the general user profile used in the application.

Implemented a sign in query in query.js telling GraphQL how a sign in attempt must be constructed and how it should handle a sign in.
The query returns a sign in model, which is defined in model.js.

Added a new sign in route to GraphQL which uses the sign in query defined in query.js.
The route is a mutation query because it changes the state of the database by saving a new refresh token to it.

Updated feedback message and http status code sent to user by the sign in logic.

Imported config to hashing.js to allow the file to use the response type variable to define message type.

Dummy data formated with prettier and user with id 1 has its password hashed to test sign in functionality.