Backend: Package Manager
Changed package manager on backend to use yarn instead of npm.

Added a yarn.lock file and deleted package-lock.json as they are in conflict with each other.

Changed the path from a dynamic path using the environment variable to a fixed file name. The path is provided to the config function of dotenv-yaml library.

Updated the package.json file with a cross-env dependency to allow the scripts to set NODE_ENV variable cross platform.
Updated all scripts to use yarn to run node commands.