Backend: Refactor Database Connection
Refactored the database connection based on the environment mode the server is started in.
The connection is established by the connectToDatabase function.
It checks what mode the server is in and connects to the database used for testing, development or production.

The mode and the file path for the database to use is stored in the config.js file which is used by connection.js.

If the server is started in any non-production modes, connectToDatabase function will call the connectToNonProductionDatabase to establish the connection.

The helper function checkEnvironmentMode checks what mode the server is started in to desife which database to connect to.

When the server starts it calles the connectToDatabase function before it goes live.