Backend: Minor Changes to Query and Test File
Instead of query.js in src/Database/Query folder initializes a database, it now imports the database from connection.js.
It leaves it to the connection.js file to define which database to connect to depending on the mode of the server.

The tests must now run connectToDatabase in its beforeAll-function to establish the database before the tests are runned or else the database will be undefined.