Backend: Refactor Dummy Data
Dummy data files are deleted from src/Database folder and moved to a new static folder named assets in the root directory of the backend folder.

The assets folder contains a database folder, with a development and test folder in it.
The development folder contains json file with dummy data used during development and manual testing. If the server runs in development mode, this is the folder it uses as database.
The test folder contains the same dummy data as the development folder, but is used and manupulated during the execution of tests.
If the server runs in test mode, this is the folder it uses as database.