Backend: Refactored Access Token
Refactored access_token.js file with a single method defining the response object.
Added check for response type in createRefreshToken function, throwing an error if refresh token is not stored.

Updated hasDataEntry function in query.js file with a check for equal value.
The missing check caused the refresh token to not be deleted from database as the function claimed it was not found.