Backend: Authenticate Token
Implemented new query for authenticating refresh tokens if the user is not logged out.

Implemented a new AuthenticateToken model for GraphQL.
The model defines the response when a user authenticate a refresh token.

New route added to route.js for GraphQL to accept requests to. The endpoint is reached at authenticateToken.
It takes a single refresh token as argument.