Backend: Authentication and Authorization Endpoint
Added new routes to GraphQL.
It now provides verifyToken and getNewToken endpoints for authorization.
The verifyToken endpoint verifies that an access token is still valid.
The getNewToken endpoint authorize the refresh token and creates a new access token which is returned to the user.
GraphQL provides a new route named signOut for authentication.
The signOut endpoint takes a refresh token as an argument and requires the access token to be set in the header of the request. It verifies the usr is logged in and deletes the refresh token from the database to sign out the user.

Added a authAccessToken query to the query.js file in the src/API/GraphQL/Query folder.
The query authorize the access token by validating that the access token is still valid or has not expired.
Added a signOut query to the same file. It sign out a user by deleting the refresh token from the database.

Implemented a new model for GraphQL to use.
The new model is a normal response model responding with a message, status code, and type of response.