Implement token verification in frontend
The access token is no set as a bearer token on every API request.
Logic for verifying and getting a new access token is implemented.
The AuthProvider component state is simplified and exposes the
isAuthenticated attribute that is used to check if the user is
authenticated every time the application starts up.