Backend: Dynamic Population of Database
Created a new folder in assets called Static which contains unmodified dummy data.

Added new folder paths to ENVIRONMENT Test and Development in config.js file.
The paths are used to check if Database for development or test mode is created, and to copy dummy data from static folder to the database folder based on the environment mode the server is started in.

Updated code in connection.js file to create Database folder in the asset folder if it does not exist. It also creates a Test or Development subfolder in the Database folder dependent on the mode the server starts in.
Added a new function which copies data from dummy_data.json file in the Static folder to the database folder of the mode the server is started in.

Added assets/Database folder to gitignore to untrack changing files in the folder.