Backend: Permission, Queries and Endpoints
Added new routes to GraphQL for get permissions,
set permission and swap permission group.

Implemented queries for getPermissions, setPermission, and swapPermission.
The getPermisssions query retrieves the available permissions an admin can update a user profile with.
The setPermission query lets an admin update the permission level of a user.
The swapPermission query lets an admin swap all users with control to experimental group, and vice versa.

Added new models for GraphQL to use.
The first new model is a PermissinEnum which is used when updating a user's permission.
The second new model is PermissionStructure, defining all permissions available in the system.
The third new model is PermissionModel defining the reponse when an admin wants to see the permission structure of the system.

The permission.js file implements an object defining the permissions available in the system, along with the functions checking a permission level, setting a permission level, and swapping permission groups on the userbase as a whole.

Added a new entry in env object representing the permission levels available in the system.