Backend: Fixed Tests
Implemented a getDatabase function to return the database from connection.js.
It secures that the database is initialized before any other function tries to access it in runtime.

Swapped out all direct refrences to database in query.js with a function retreiving the database from connection.js.

The access token test file uses the user table and refresh token table name from the config file instead of a hard coded value.

Removed the direct access to the test dummy_data.json file as it made jest crash because a file was imported before initialization.
Instead the beforeAll function retrieves the necessary data through the query method of the database, which secures that the dummy_data file is initialized before the test tries to access it.

Added the connectToDatabase function to hashing.js to make sure the beforeAll function initialize the database before the test runs.

Added an afterAll clean up function to all tests resetting the data in the test database to its initial state to avoid bugs and unstable results when running test script several times in a row.