Backend: Fixed Logic Flaw In Connection File
Fixed a logic flaw in connection.js in resetTestData, where the database reference was not updated with a new reference causing the tests to run half and half with correct and uncorrect data.

Added resetTestData function in beforeAll in all test files to make sure the database is reset before each run.