Backend: Account Endpoints and Account Activation
Changed name of field in dummy data from name to username.
Added valid usernames pre-authorized to create an account on the application.

Added password.js file in src/Security folder to internal.js.

Added new routes to GraphQL.
The route verifyUsername verifies if a username entered during registration is pre-authorized to create an account.
The activateAccount route activates and stores the account the user has been given to the database.

Refactored GraphQL models to use camelCase instead of snake_case. All fields with id as name has lower case, while fields ending on id have ID as upper case.
The field name no longer exist on the user model and has been renamed to username.

Implemented verifyUser query in query.js for GraphQL, which checks if a username is pre-authorized to register an account.
Returns a status 200 and a verification message if the username is verified.

Implements activateAccount query in query.js for GraphQL, which activates the pre-authorized account by creating a profile and store it to database.

Renamed name field in all queries to username.

Implemented signUp functionality in sign_up.js.
It checks if a user exist before it is created.
A user is created by hashing their defined password, creating a profile, updating the profile with the new data, storing the user in database, and signing the user in.

Added a new environment variable to the config.js file specifying the table of the pre-authorized users who can register to the application.

Added a new file to src/Security folder which takes care of password security.