Refactor login and logout functionality
The API login logic is now moved from the AuthProvider component to the
LoginScreen component. The LoginScreen uses hooks from the Apollo
library to query the backend. The AuthProvider component exposes
functions to persist the authenticated user and it's tokens. Also, the
logout logic is moved into a shared LogoutButton component using the
same tactic. These changes greatly reduces the complexity of the
AuthProvider component.