Backend: Fixed Bug Overwriting Users in Database
Fixed a bug which caused users with the latest id in user table to be overwritten.
The bug was caused by the nextID function in query.js because it returned the ID as the length of the table and not the latest user's ID pluss 1.
When a user was deleted the from the table anywhere except the last position, the ID of the last user would be for example 70, but the length would be 69, giving the next user created the value 70 and overwrite the last user.