Backend: Added State to Profile, Renamed Variable
Added an object defining which state a profile can have.
The available states are:
- deactivated: The profile is not in use yet, need to reset password, or must be unavailable during swapping of permission group the profile should belong to.
- active: The username has been verified is a part of the research experiment and the participant in the research experiment has created a password for the profile, activating the profile.
- suspended: The profile is suspended due to following reasons: 1) Participant has resigned to participate or withdrawn their concent to contribute to the research anymore, 2) Has violated the agreement between the researchers and the participant, 3) If non-authorized activity is detected.

Renamed variable key to functionKey in profile.js.

Added a state variable to the profile object.
Added a has state function to the profile object, which can be used to check a state during sign in or usage of other endpoints requiring the user to have an active state.