Backend: Password Validation, andFeedback
Added an object defining generic feedback messages to the user if a password is not valid.

Implemented helper functions to:
- validate length of password
- check if password contains valid capital letters, lower case letters, numbers, and special characters.
- concatinate error mesages

Implemented functions to validate a password, check if the result of the validation is valid, get errors found by the validation, and generate a single message from all the errors found.