Backend: Update verifyUsername Query
Added new logic to the verifyUserQuery in query.js in src/GraphQL/Query folder.
The new logic checks if a user exist when a participant tries to verify the username, and if it does, checks if the profile is in an active or suspended state. Active and suspended state should prevent the user from activating the profile again.