Fix error resolving module @apollo/client
Prior to this fix, when trying to launch the app on Android, the Android
bundling fails. Metro was not able to resolve the @apollo/client module,
due to Apollo using the `.cjs` file extension, which Metro don't
understand. Fixed by configuring the Metro config resolver to accept
this file extension.

Closes #66