Backend: Implemented Change Password
Implemented change password functionality in profile.js.
It first checks if passwords match, then the format is validated, and finished with hashing and updating the user's password.

Implemented a change password query, responding with a message, status and type.
It checks if a user is logged in with an access token, then validates if the logged in user is changeing their own password, then retrieving the user's data from database and checks if it exists.
If user exist the profile is updated with information in the database before the password is changed, and user updated in database.