Backend: Added getFilter Function to All Providers
The getFilter function in the json_provider.js and firebase_provider.js returns the format of the filter they use to filter data in database.
The function is used by the query.js file in src/Database/Query folder to set up the filter to use based on the provider and return it to the callee.