Backend: Minor Logic Flaw Fixes
When checking permission the permission must be the correct level to get access.
previous logic allowed experimental group to access all content of control group which should not be allowed.
The flaw could potensially made a participant in the experimental group get it's app stripped for content as it was a part of the control group, which should not be the case.

The nextID functionin json_provider.js used resolve instead of return when it did not use promises, causing a runtime error.

The firebase_provider.js file has updated the key names in updateData function, as new names were given when refactoring the getFilter function in query.js file.

The getFilter function in query.js now returns the correct filter to use depending on if the database is a json file or Firebase.
It does so after changing the if statement to check the object's hasOwnProperty.