Backend: Updated Access Token File
Updated the createAccessToken function to use username instead of name.

The storeRefreshTokenInDatabase function now handles refresh tokens stored as a single element in a list and as an object.

The authenticateRefreshToken function now checks if a refresh token exist in either a list or an object.

Updated the access token test file to test according to the changes of how an access token is stored.