Backend: Implemented Category Features
Moved functions to other places in the file to satisfy ESLint.

Implemented an updateCategory function updating a category object's data.

Implemented the addChallenge function to add a challenge to the category if it does not already exist.

Implemented the removeChallenge function removing a challenge if it exist in the category.

Implemented a convertToStoredObject function returning the format a category should be stored as in the database.

Implemented a convertToResponseObject function returning the format a catgory should have in a response back to the client.

Implemented a restoreObject function retrieving all sub parts a category consists of from the database and re-asembles the category object to be used in the application.

Implemented a saveData function saving a category to the database and returns the object it saved.