Backend: Implemented Content Features
Moved functions around in file to satisfy ESLint.

Renamed name key for all functions to functionKey instead, as the variable name key collided with key in the updateContent function.

Implemented a convertToStoredObject function returning the format of how a content object should be stored in database.

Implemented a convertToResponseObject function returning the format the content object should have when a response is sent to back to the client.

Implemented a saveData function saving a content object to the database and returning the object it saved.