Backend: Implemented Category Queries
Implemented a getAllCategories query retreiving all categories available in the application.
It uses the AllCategoriesResponseModel to return it's result back to client.

Implemented a createCategory query allowing an administrator of the application to create a new category with content.
Content is title, image, and a description.
The category and the content is stored to database and returned as an object to the client in the response.