Use environment variable to set API endpoint
When running against backend on localhost, the actual IP address is
needed for it to work on Android. For the frontend to be able to connect
with the backend, the API_ENDPOINT environment variable needs to be set.
The full URL must be set (including both scheme and port) either with

    export API_ENDPOINT="http://192.168.1.42:5000/graphql"
    yarn start

or by putting the variable in a .env file