Use the GOOGLE_APPLICATION_CREDENTIALS env var
As documented by Google, it is possible to use the
GOOGLE_APPLICATION_CREDENTIALS environment variable to set the Google
Service Account credentials when initializing the Firebase Admin SDK.
Thus it's no need to introduce our own variable for this purpose.

Also updated the README.md file with a section for running the backend
in production mode.