Backend: Updated Sign Up
Sign up function is updated to reassign validUsername and userObject after promise is resolved.

The get nextID function now uses the nextID function of the provider, and does no longer need to return a promise.

As nextID no longer return a promise, the updateNewUser function does no longer need to return a promise as well.