Backend: Implemented Functions for Challenge
The updte method for challenge checks for null and if input is an object.

Removed createObjectTemplate function and replaced it with a function in utility.js intstead for more reuse.

Implemented a restoreObject function in challenge.js.
The function populates the challenge object and all its sub-objects with data.

Implemented the deleteChallenge function, which removes the challenge from a category and deletes the challenge from the database.

Implemented an addReflection function allowing a user to specify which type of reflection they want to use in a challenge.
This could be set when a user creates a challenge.

Updated ids in object to be a string instead of an int.