Backend: Updated Implementation of Category
Added a check for null and object to the updateCategory function.

Fixed a logic flaw when adding challenge to a category.
The function now checks if the challenge to be added to a category exist in the list before.
If it is not in list, it will be added.

Refactored restoreObject of category.js to reuse code in utility.js file as the code is similar to the one of restoreObject in challenge.js.
The restore object of category also restores challenge objects based on their ids.