Backend: Implemented New Challenge Queries
Implemented a query retrieveing all challenges from the database.

Implemented a query creating a challenge if the user is an admin.
By specifying a category id the challenge will be added to the category if it exist.
The challenge gets stored and is returned to the user in the response.