Backend: GraphQL Model for Click Stream
Implemented a GraphQL model for creation of click streams.

Added a meta data model to be used when collecting click treams.
The meta data registers data about previous screen a user has visited, the challenge a user is visiting, and what source in app the user has triggered. A source referes here to a button, a link, etc. Navigations are not included as it is considered an event.