Backend: New Click Stream Endpoint for GraphQL
Added a new endpoint to create a new click stream for GraphQL.
The endpoint takes as input a createClickStream model and verifies the user to collect the click stream for is logged in.
It then stores the click stream of the user to the database.

The new endpoint is a mutation endpoint as it is changing the state of the database.
The endpoint is named createClickStream.