Backend: Delete Click Stream Implementation
Modified the code of deleteClickStream functionality as clickStreamItem is no longer needed because of new requirements to the mplementation.
The function deleteClickStream no longer need a clickStream object as argument because it no longer uses a clickStreamItem object to store click data. This is because the click stream object can handle it without much complexity added.
As clickStream is removed as input, the function does no longer need to check for null or if it is an object.
The function does no longer need to chain deletion down to clickStreamItems as the clickStream item now takes care of the logic the clickStreamItem was supposed to handle.
The function calls deleteData to delete the clickStream from the database and returns a 200 response on success.