Backend: New Query for Deleting Click Stream
Implemented a query for deleting click streams by id.
The query takes an input of a click stream id, checks if the user is logged in and has permission of admin, retrieves the click stream with the input id, restores the click stream object, and deletes the click stream from database.
It responds with a 200 status code on success.