Backend: Modified code, New functions for utility
Modified the getDataFromDatabaseByFilter function to have a key parameter as well as value and table.

Added sessionToken as field to the stored click stream object format.

Implemented a getClickStreamResponseObject, returning the format a click stream should have in a response.

Added click stream as case to the function converting an object to a response object.