Backend: New Click Stream Query and Modification
Implemented a new query for GraphQL, retrieving all click streams stored in the database.
The query requires that the user is logged in and has permission of admin.
It retrieves and restores all click  streams to objects and sends their data as response.

The endpoint is used to get the data needed to analyse how users have interacted with the application.
Because of restrictions for privacy data, the application must capture such data because third party liberaries cannot be used.

Modified the code of createClickStream query to take as input a session token, which is the user's refresh token, as a user gets a new refresh token very time they log in and out.
The refresh token must exist, and if a click stream is registered on that session before, it will continue to log data to that session. Otherwise a new click stream session is logged.