Backend: Refactor and implemented Reward
Refactored reward.js by moving code in file to satisfy ESLint.
Updated variable name for key in reward object to functionKey instead of key, as key collides with another variable in the same file.

Implemented delete reward function.
It takes as input a challenge containing the reward object to be deleted and checks if the challenge is not null or not an object.
The reward is deleted from the database, and the challenge updates it's reward object with an empty reward.