Backend: Reward Stored and Response Object
Implemented a function in utility.js returning the format of how the reward object should be stored in database.
Added a case for reward when converting the reward object into a stored object.

Implemented a function in utility.js returning the format of the reward response object which should be sent in the response back to the user.
Added a case for reward when converting the reward object into a response object.