Backend: Refactor Challenge Query
Refactored out code from create challenge query to a new utility file located in src/API/GraphQL/Query/Util.
The refactored code in utility creates and returns a challenge, its content, and the reward.
The code saving the content, challenge, reward, and category is also refactored out to the same file.

The create challenge query now also takes as input a reward, and creates a reward object when a challenge is created.