Backend: Implemented and Refactor Reflection
Refactored code in reflection.js by moving code around to satisfy ESLint.
Renamed key to functionKey to be consistent between files.

Implemented the setTitle function, which checks if input is valid string and updates the title.
Implemented the setSolution function, checking if a solution is a string and updates the solution.
Implemented addChoice function adding a choice if the reflection type is argument.
Implemented removeChoice function, removing a choice from the list of choices to create an argument.
Implemented the deleteReflection function. It takes as input a challenge to delete the reflection from and checks if the challenge is an object.
The reflection is then deleted and the challenge is updated with an empty reflection of the same type.