Backend: Challenge Query Supports Reflection
The create challenge query is modified to take a reflection as input .
The challenge is now created with the reflectionType argument telling the challenge waht reflection type it should create.
The reflection object is created and initialized with data before it is saved to the database.

The challenge creator now creates a challenge with the reflection type.

The save challenge function now saves a reflection to database.

The create reflection function gives the reflection an id and populates its data. If the reflection is an argument the choices are populated as well.