Backend: GraphQL Reflection Model
Implemented a reflection input model defining how an input for a reflection should be sent in a request. The choice field need only input if the reflection type is set to an argument when creating a challenge.
The reflection type enum model allows the input to define if a reflection is a regular reflection or an argument by giving the field an input of "reflection" or "argument" in string format which is then converted by GraphQL to a boolean used by the backend.