Backend: Implemented Activity
Updated empty data function for activity to have an ID as string, a type defining if activity is in app or have to be done outside of application, a description summarizing how to perform the activity step by step.

Implemented set id function updating the id of the activity.

Implemented a set type function updating the type of an activity.

Implemented a set description updating the description of an activity.

Implemented add and remove function adding or removing an element to or from the input list.

Implemented a delete activity function deleting an activity from a challenge and in database.