Backend: GraphQL Query for Deleting Activity
The query for deleting an activity takes as input an activity ID for the activity to delete.
It checks if user is logged in and has permission of administrator.
All challenges to delete an activity from is retrieved by the activity id and lopped through for deletion of the activity.