Backend: Updated Dummy Data with Real Challenge
Dummy data now contains a real challenge related to the category molecular movements.

Added challenges with data to dummy data.

Modified code of emulator initialization to use the ids assigned to the objects instead of creating new once.

When adding new challenges and categories use the generateIDs function to get valid ids from Firestore to use when creating a new content to add to dummy data.