Backend: Refactor Utility
Removed unused functions after refactoring.

Reduced the complexity of restore challenges function as the restore object function of a challenge got its complexety reduced and no longer returns a promise.

The function storing a challenge no longer stores ID for sub objects, but stores the data of the object inside the fields.