Backend: Refactor Core Due to New Firestore Implem
All sub objects of challenge and category is now located in thier documents, causing the following changes:

Deleted file for activity query as delete activity is no longer needed. The new structure of challenge in database creates a logic where an activity is created and deleted along with a challenge and is therefore not needed.
The same argument holds for reward query file.

As all database methods are done by a new firestore provider, save data method in each object is no longer needed and is therefore deleted as a function from objects and from utility file.

All queries and functions have been refactored to use the provider instead.
This change caused a restructure of code to make it work as before.

The utility file have been reduced in complexity as a lot of the functions are no longer needed.
The functions for restoring a challenge object and category object and their stored counterpart have been moved into their respective files, meaning they no longer reside in utility.